Meals refactored to MVC style and all tests passing
